feat(PilotManager): use caseId and add fill_data and add trace_id
